fantasy-draft-agent/requirements.txt
any-agent[openai]
python-dotenv
pydantic
gradio
# Optional: For enhanced console demos
# rich
